fix: prevent formatting of files in _build
If working with an application (instead of a release) the
modification would format files in _build, too.  Prevent
that by formatting the directories src, test, and include
explicitly.

Using rebar3_dir:all_src_dirs/1 is way more trickier as
this function reads the values from the rebar configuration
which does not include sane defaults or is dependent on
profiles.  We just want to format everything.